map_navigator.py file added
